pyproject.toml
[project]
name = "mcp-server-box"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "box-ai-agents-toolkit",
    "cryptography>=45.0.4",
    "fastmcp",
    "langchain",
    "langchain-openai",
    "pydantic",
    "pyjwt>=2.10.1",
    "python-dotenv",
]

[dependency-groups]
dev = ["pytest>=8.3.5", "pytest-asyncio>=0.26.0", "pytest-cov>=6.1.0"]
